requirements.txt
langchain
langchain-google-genai
huggingface_hub
python-dotenv
streamlit